Move CoinToCointContext to correct package
